requirements.txt
fastapi
uvicorn
pydantic
openai
jinja2
pytz
requests
playwright
pyyaml
imageio-ffmpeg
